refactor: standardize kernel file naming to <op>_<dtype>_<role>
Rename FlyDSL kernel files to the op-first <op>_<dtype>_kernel.py pattern,
matching the already-consistent gemm_bf16_kernel.py / gemm_fp8_kernel.py and
the PyTorch frontend <op>_<dtype>_impl.py layout:

  flydsl/gemm/mxfp4_gemm_kernel.py            -> gemm/gemm_mxfp4_kernel.py
  flydsl/gemm/mxfp8_gemm_kernel.py            -> gemm/gemm_mxfp8_kernel.py
  flydsl/grouped_gemm/gemm_fp8_grouped_kernel.py -> grouped_gemm/grouped_gemm_fp8_kernel.py
  flydsl/grouped_gemm/mxfp4_grouped_kernel.py    -> grouped_gemm/grouped_gemm_mxfp4_kernel.py
  flydsl/grouped_gemm/mxfp8_grouped_kernel.py    -> grouped_gemm/grouped_gemm_mxfp8_kernel.py

Rename the CK WS grouped-gemm heuristic to carry its op prefix:
  pytorch/kernels/grouped_gemm/ws_ck_heuristic.py -> grouped_gemm_ck_ws_heuristic.py
  (+ its test)

Update all import references (flydsl cross-imports, pytorch frontends, tests)
and the LICENSE file listing. Document the convention in CONTRIBUTING.md,
including why the PyTorch frontend keeps dtype-family names (fp8/fp4 dispatch
tensorwise AND mx kernels) while the FlyDSL layer names concrete formats.

No functional changes; py_compile passes for all touched files.
